services/deepseek_service/requirements.txt
# Core API dependencies
fastapi
uvicorn[standard]
pydantic
python-multipart

# DeepSeek-OCR dependencies (except torch/flash-attn)
transformers==4.46.3
tokenizers>=0.20.3
accelerate>=0.26.0
einops
addict
easydict
matplotlib
Pillow
ninja
packaging

# NOTE: torch, torchvision, flash-attn must be installed via special wheels (see Dockerfile)
